+++ eugene/datasets/_datasets.py
@@ -1,57 +1,18 @@
 from pathlib import Path
 from typing import Optional
 import warnings
-import wget, os, gzip, io
 
 import numpy as np
 import pandas as pd
 
 from .._compat import Literal
 from .._settings import settings
-from ._utils import check_datasetdir_exists
+from ._utils import check_datasetdir_exists, try_download_urls
 
-from ..dataloading._io import read
+from ..dataloading._io import read, read_csv
 HERE = Path(__file__).parent
 pkg_resources = None
-
-def try_download_urls(data_idxs: list, url_list: list, ds_name: str, is_gz: bool = False) -> list:
-    paths = []
-    for i in data_idxs:
-        csv_name = os.path.basename(url_list[i]).split(".")[0] + ".csv"
-        if not os.path.exists(os.path.join(HERE.parent, settings.datasetdir, ds_name, csv_name)):
-            ds_path = os.path.join(HERE.parent, settings.datasetdir, ds_name)
-            if not os.path.isdir(ds_path):
-                print(f"Path {ds_path} does not exist, creating new folder.")
-                os.mkdir(ds_path)
-
-            print(f"Downloading {ds_name} {os.path.basename(url_list[i])} to {ds_path}...")
-            path = wget.download(url_list[i], os.path.relpath(ds_path))
-            print(f"Finished downloading {os.path.basename(url_list[i])}")
-
-            if is_gz:
-                with gzip.open(path) as gz:
-                    with io.TextIOWrapper(gz, encoding="utf-8") as file:
-                        file = pd.read_csv(file, delimiter=r"\t", engine="python")
-
-                        if ds_name == "deBoer20":
-                            file = deBoerCleanup(file, i)
-
-                        save_path = os.path.join(ds_path, csv_name)
-                        file.to_csv(save_path, index = False)
-                        print(f"Saved csv file to {save_path}")
-                        paths.append(save_path)
-                os.remove(os.path.join(ds_path, os.path.basename(url_list[i])))
-            else:
-                # If file is not packed, same logic but without gzip
-                pass
-        else:
-            print(f"Dataset {ds_name} {csv_name} has already been dowloaded.")
-            paths.append(os.path.join(HERE.parent, settings.datasetdir, ds_name, csv_name))
-    return paths
-
-def deBoerCleanup(file: pd.DataFrame, index: int) -> pd.DataFrame:
-    return file
-
+        
 def get_dataset_info():
     """Return DataFrame with info about builtin datasets.
     Returns
@@ -111,10 +72,18 @@ def deBoer20(datasets: list, **kwargs: dict) -> pd.DataFrame:
         "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_20160609_average_promoter_ELs_per_seq_pTpA_ALL.shuffled.txt.gz",
         "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_20161024_average_promoter_ELs_per_seq_3p1E7_Gal_ALL.shuffled.txt.gz",
         "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_20161024_average_promoter_ELs_per_seq_3p1E7_Gly_ALL.shuffled.txt.gz",
+        "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_20170811_average_promoter_ELs_per_seq_OLS_Glu_goodCores_ALL.txt.gz",
         "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_20180808_processed_Native80_and_N80_spikein.txt.gz",
         "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE104nnn/GSE104878/suppl/GSE104878_pTpA_random_design_tiling_etc_YPD_expression.txt.gz"
     ]
 
+    if type(datasets) is int:
+        datasets = [datasets]
+
     paths = try_download_urls(datasets, urls_list, "deBoer20", is_gz = True)
 
-    print(paths)
+    seq_col="SEQ"
+    target_col="TARGET"
+
+    data = read_csv(paths, seq_col=seq_col, target_col=target_col, sep=",", col_names=[seq_col,target_col], **kwargs)
+    return data
